modify: modifies code in Position.do_move_at method.

+++ pyrev/pyrev.pyx
@@ -395,7 +395,7 @@ cdef class Position:
 			self.do_pass()
 
 		cdef uint64_t moves = self.__bb.calc_player_mobility()
-		cdef uint64_t c = coord
+		cdef int32_t c = coord
 		cdef uint64_t x = 1 << c
 		print(c)
 		print(x)
